no file case

+++ app/controllers/items_controller.rb
@@ -250,14 +250,16 @@ class ItemsController < ApplicationController
   end
 
   def delete_items
-    path = params[:csv_file].tempfile.path
-    open(path, 'r:cp932:utf-8', undef: :replace) do |f|
-      csv = CSV.new(f, :headers => :first_row)
-      csv.each do |row|
-        next if row.header_row?
-        code = row.fields
-        item = Item.find_by(user_id: current_user.id, code: code)
-        item.destroy if item.present?
+    if params[:csv_file]
+      path = params[:csv_file].tempfile.path
+      open(path, 'r:cp932:utf-8', undef: :replace) do |f|
+        csv = CSV.new(f, :headers => :first_row)
+        csv.each do |row|
+          next if row.header_row?
+          code = row.fields
+          item = Item.find_by(user_id: current_user.id, code: code)
+          item.destroy if item.present?
+        end
       end
     end
